Harden translation catalog loading against switch races, load failures, and nondeterministic precedence
